Restructure project and update imports
